AIML/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning & Deep Learning
torch==2.2.0
torchvision==0.17.0
open-clip-torch==2.24.0
sentence-transformers==2.3.1
scikit-learn==1.3.2
joblib==1.3.2

# Data Processing
pandas==2.1.3
numpy==1.26.4
tldextract==5.1.1

# Image Processing
Pillow==10.1.0
imagehash==4.3.1
opencv-python-headless==4.8.1.78

# ChromaDB Client
chromadb>=1.1,<1.2

# Kafka Client
kafka-python==2.0.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Logging
python-json-logger==2.0.7
